+++ Mods ChoGGi/Pause On Load/Code/Script.lua
@@ -1,17 +1,44 @@
 -- See LICENSE for terms
 
+local mod_NewPause
+local mod_LoadPause
+
+-- fired when settings are changed/init
+local function ModOptions()
+	mod_NewPause = CurrentModOptions:GetProperty("NewPause")
+	mod_LoadPause = CurrentModOptions:GetProperty("LoadPause")
+end
+
+-- load default/saved settings
+OnMsg.ModsReloaded = ModOptions
+
+-- fired when option is changed
+function OnMsg.ApplyModOptions(id)
+	if id ~= CurrentModId then
+		return
+	end
+
+	ModOptions()
+end
+
+
 function OnMsg.CityStart()
+	if not mod_NewPause then
+		return
+	end
 	CreateRealTimeThread(function()
-		-- just throwing them out there
-		WaitMsg("MarsResume")
-		Sleep(1000)
+		WaitMsg("MessageBoxClosed")
+		Sleep(100)
 		SetGameSpeedState("pause")
 	end)
 end
 
 function OnMsg.LoadGame()
+	if not mod_LoadPause then
+		return
+	end
 	CreateRealTimeThread(function()
-		WaitMsg("MarsResume")
+		Sleep(100)
 		SetGameSpeedState("pause")
 	end)
 end