+++ Mods ChoGGi/Pause On Load/items.lua
@@ -1,0 +1,12 @@
+return {
+	PlaceObj("ModItemOptionToggle", {
+		"name", "NewPause",
+		"DisplayName", T(302535920011076, "Pause on New Game"),
+		"DefaultValue", true,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "LoadPause",
+		"DisplayName", T(302535920011565, "Pause on Load Game"),
+		"DefaultValue", true,
+	}),
+}