+++ Mods ChoGGi/Solylent Glutton Provides Double/metadata.lua
@@ -1,0 +1,20 @@
+return PlaceObj("ModDef", {
+	"title", "Solylent Glutton Provides Double",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_SolylentGluttonProvidesDouble",
+	"steam_id", "1907034997",
+	"pops_any_uuid", "32c54f4c-6008-4644-882c-5cc8403a08a9",
+	"author", "ChoGGi",
+	"lua_revision", 249143,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[If you have the solylent green breakthrough; gluttons will provide double the soylent.
+
+
+Not exactly requested by Monovigma...]],
+})